backend/requirements.txt
# Core Application
fastapi==0.111.0
uvicorn[standard]==0.29.0
gunicorn==20.1.0
python-dotenv==1.0.1
python-multipart==0.0.9

# Database & ORM
sqlalchemy==2.0.29
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.27.0

# Caching
redis==4.5.0

# Security & Auth
pydantic[email]==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]>=3.3.0
passlib[bcrypt]==1.7.4

# Task Queue
celery==5.2.0

# Cloud (Optional)
# cloud-sql-python-connector[pg8000]==1.2.0

# Development & Testing
pytest==7.2.0
pytest-asyncio==0.21.0
httpx==0.23.0

# Logging
structlog==22.3.0
